engine_image_cls/requirements.txt
matplotlib==3.5.1
numpy==1.21.6
opencv-python-headless==4.5.5.64
Pillow==9.1.0
scipy==1.7.3
torch==1.7.1
torchvision==0.8.2
timm==0.5.4
tqdm==4.64.0
albumentations==1.1.0
pandas==1.3.5
requests
thop
